refactor(epicenter): delete epicenter/ folder and move iterActions to workspace/
Move iterActions and ActionInfo from core/epicenter/ to core/workspace/
since the epicenter folder was nearly empty after the API consolidation.
The workspace/ folder now contains all client-related code including
action iteration utilities.
