--- packages/epicenter/src/core/epicenter/index.ts
@@ -1,8 +1,0 @@
-/**
- * Default epicenter exports (Node.js).
- *
- * Internal code (CLI, server, tests) imports from this file.
- * These are all Node-only contexts, so we re-export the node version.
- */
-
-export * from './index.node';
